requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pymongo==4.6.1
pytest==7.4.3
python-dotenv==1.0.0
Werkzeug==3.0.1
